chore(test): move tests to the source dir (#358)
Rename test/ to tests/ to contain only integration tests and common test
utilities (`common/`). Module unit tests should be created next to the
module source with the name `<module name>_test.py`.

This makes them easy to discover and obvious when unit tests are
missing.

Added README to describe the test structure.
