+++ .gitignore
@@ -91,8 +91,9 @@ dmypy.json
 !.vscode/extensions.json
 
 # Generated files
-onnxscript/test/models/testoutputs/*
+onnxscript/tests/models/testoutputs/*
 docs/auto_examples/*
-onnxscript/test/mylib.onnxlib
-onnxscript/test/onnx_backend_test_code
+onnxscript/tests/mylib.onnxlib
+onnxscript/tests/onnx_backend_test_code
 *.onnxlib
+*.onnx
