+++ onnxscript/tests/README.md
@@ -1,0 +1,5 @@
+# tests
+
+This directory contains integration tests and common test utilities (`common/`).
+Module unit tests should be created next to the module source with the name `<module name>_test.py`.
+This makes them easy to discover and obvious when unit tests are missing.
