+++ onnxscript/tests/cli_test.py
