+++ onnxscript/tests/functions/onnxfns_test.py
@@ -7,8 +7,8 @@ import unittest
 
 import onnx
 
-from onnxscript.test.common import onnx_script_test_case
-from onnxscript.test.models import onnxfns1
+from onnxscript.tests.common import onnx_script_test_case
+from onnxscript.tests.models import onnxfns1
 
 
 class TestOnnxFns(onnx_script_test_case.OnnxScriptTestCase):
