+++ onnxscript/tests/models/identity.py
@@ -14,19 +14,23 @@ from onnxscript.onnx_types import BOOL, FLOAT, INT64
 def id1(A: FLOAT[...]) -> FLOAT[...]:
     return A  # treat as op.Identity(A)
 
+
 @script()
 def id1_expanded(A: FLOAT[...]) -> FLOAT[...]:
     return op.Identity(A)
+
 
 @script(default_opset=op)
 def id2(A: FLOAT[...]) -> FLOAT[...]:
     B = A
     return B  # treat as op.Identity(B) == op.Identity(A)
 
+
 @script()
 def id2_expanded(A: FLOAT[...]) -> FLOAT[...]:
     B = A
     return op.Identity(B)
+
 
 @script()
 def control_flow_id1(A: FLOAT[...], flag: BOOL) -> FLOAT[...]:
@@ -36,6 +40,7 @@ def control_flow_id1(A: FLOAT[...], flag: BOOL) -> FLOAT[...]:
         y = op.Abs(A)
     return y
 
+
 @script()
 def control_flow_id1_expanded(A: FLOAT[...], flag: BOOL) -> FLOAT[...]:
     if flag:
@@ -44,6 +49,7 @@ def control_flow_id1_expanded(A: FLOAT[...], flag: BOOL) -> FLOAT[...]:
         y = op.Abs(A)
     return y
 
+
 @script()
 def loop_id(A: FLOAT[...], N: INT64) -> FLOAT[...]:
     B = op.Identity(A)
@@ -51,6 +57,7 @@ def loop_id(A: FLOAT[...], N: INT64) -> FLOAT[...]:
         B = A  # treat as op.Identity(A)
         A = A + 1
     return A + B
+
 
 @script()
 def loop_id_expanded(A: FLOAT[...], N: INT64) -> FLOAT[...]:
