+++ onnxscript/tests/models/opt_output.py
@@ -1,5 +1,3 @@
-from tkinter import Y
-
 from onnxscript import script
 from onnxscript.onnx_opset import opset15 as op
