+++ pyproject.toml
@@ -32,7 +32,8 @@ dependencies = ["numpy", "protobuf<4", "onnx", "typing_extensions"]
 version = {attr = "onnxscript.__version__"}
 
 [tool.setuptools.packages.find]
-exclude = ["onnxscript.test.*"]
+include = ["onnxscript*"]
+# Tests are by default excluded: https://setuptools.pypa.io/en/latest/userguide/package_discovery.html
 
 [tool.setuptools.package-data]
 onnxscript = ["py.typed"]
@@ -77,21 +78,21 @@ disallow_untyped_defs = true
 [[tool.mypy.overrides]]
 module = [
   "setup",
-  "onnxscript.test.models.*",
-  "onnxscript.test.onnx_backend_test_code.*",
+  "onnxscript.tests.models.*",
+  "onnxscript.tests.onnx_backend_test_code.*",
 ]
 ignore_errors = true
 
 [tool.black]
 target-version = ["py37", "py38", "py39", "py310"]
 # Black's extend-exclude needs to be a regex string
-extend-exclude = "/onnxscript/test/models|/onnxscript/test/onnx_backend_test_code"
+extend-exclude = "/onnxscript/tests/models|/onnxscript/tests/onnx_backend_test_code"
 line-length = 95
 
 [tool.isort]
 profile = "black"
 extend_skip_glob = [
-  "onnxscript/test/onnx_backend_test_code/*.py",
+  "onnxscript/tests/onnx_backend_test_code/*.py",
 ]
 
 [tool.pylint.messages_control]
